Updated app.js with Coursera, Udemy, Youtube, Khan Academy, modifed identation for readability, fixed asynchronous results handling, also can handled empty selection submission (0 option checked)
